+++ src/js/disability-benefits/686/containers/Form686App.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import FormApp from '../../common/schemaform/containers/FormApp';
+import FormApp from '../../../common/schemaform/containers/FormApp';
 import formConfig from '../config/form';
 
 export default function Form686App({ location, children }) {
