+++ src/js/disability-benefits/686/reducers/index.js
@@ -1,5 +1,5 @@
 import formConfig from '../config/form';
-import createSchemaFormReducer from '../../common/schemaform/state';
+import createSchemaFormReducer from '../../../common/schemaform/state';
 
 export default {
   form: createSchemaFormReducer(formConfig)
